requirements.txt
tensorflow-gpu==1.13.2
keras==2.2.4
matplotlib
numpy
sklearn